[fix] validate and name
