+++ controllers/todoControllers.js
@@ -1,4 +1,5 @@
 const todoService = require("../service/todoService");
+const { validateTodoInput } = require("../utils/validation");
 
 exports.getAllTodos = async (req, res) => {
   try {
@@ -15,8 +16,8 @@ exports.getPaginateTodo = async (req, res) => {
     page = Number(page);
     limit = Number(limit);
     const result = await todoService.getPaginateTodo(page, limit);
-    if (!result.status || result.status !== 200) {
-      return res.status(result.status || 400).json({ message: result.error });
+    if (result.status !== 200) {
+      return res.status(result.status).json({ message: result.message });
     }
     res.status(200).json(result);
   } catch (err) {
@@ -27,17 +28,28 @@ exports.getPaginateTodo = async (req, res) => {
 exports.createTodo = async (req, res) => {
   try {
     const { content, description, type_id, isDone } = req.body;
+
+    const validation = validateTodoInput({
+      content,
+      description,
+      type_id,
+      isDone,
+    });
+    if (!validation.valid) {
+      return res.status(400).json({ message: validation.message });
+    }
+
     const result = await todoService.createTodo(
       content,
       description,
       type_id,
       isDone
     );
-    if (!result.status || result.status !== 201) {
-      return res
-        .status(result.status || 400)
-        .json({ message: result.error || result.message });
+
+    if (result.status !== 201) {
+      return res.status(result.status).json({ message: result.message });
     }
+
     res.status(201).json({ message: result.message });
   } catch (err) {
     res.status(500).json({ message: "Server error" });
@@ -48,6 +60,17 @@ exports.updateTodo = async (req, res) => {
   try {
     const { id } = req.params;
     const { content, description, type_id, isDone } = req.body;
+
+    const validation = validateTodoInput({
+      content,
+      description,
+      type_id,
+      isDone,
+    });
+    if (!validation.valid) {
+      return res.status(400).json({ message: validation.message });
+    }
+
     const result = await todoService.updateTodo(
       id,
       content,
@@ -55,11 +78,11 @@ exports.updateTodo = async (req, res) => {
       type_id,
       isDone
     );
-    if (!result.status || result.status !== 200) {
-      return res
-        .status(result.status || 400)
-        .json({ message: result.error || result.message });
+
+    if (result.status !== 200) {
+      return res.status(result.status).json({ message: result.message });
     }
+
     res.status(200).json({ message: result.message });
   } catch (err) {
     res.status(500).json({ message: "Server error" });
@@ -70,11 +93,11 @@ exports.deleteTodo = async (req, res) => {
   try {
     const { id } = req.params;
     const result = await todoService.deleteTodo(id);
-    if (!result.status || result.status !== 200) {
-      return res
-        .status(result.status || 400)
-        .json({ message: result.error || result.message });
+
+    if (result.status !== 200) {
+      return res.status(result.status).json({ message: result.message });
     }
+
     res.status(200).json({ message: result.message });
   } catch (err) {
     res.status(500).json({ message: "Server error" });
