+++ service/todoService.js
@@ -1,32 +1,5 @@
 const todoRepo = require("../repository/todoRepository");
-const { todoSchema } = require("../models/todoModels");
-
-function validateTodoInput(data) {
-  for (const key in todoSchema) {
-    const rule = todoSchema[key];
-    const value = data[key];
-    if (rule.type === "string" && value && typeof value !== "string") {
-      return `${key} must be a string`;
-    }
-    if (
-      rule.type === "number" &&
-      value !== undefined &&
-      value !== null &&
-      isNaN(Number(value))
-    ) {
-      return `${key} must be a number`;
-    }
-    if (
-      key === "isDone" &&
-      value !== undefined &&
-      value !== null &&
-      !(value === 0 || value === 1)
-    ) {
-      return "isDone must be 0 or 1";
-    }
-  }
-  return null;
-}
+const { validateTodoInput } = require("../utils/validation");
 
 exports.getAllTodos = async () => {
   return await todoRepo.getAllTodos();
@@ -47,31 +20,41 @@ exports.getPaginateTodo = async (page, limit) => {
 };
 
 exports.createTodo = async (content, description, type_id, isDone) => {
-  const error = validateTodoInput({ content, description, type_id, isDone });
-  if (error) return { status: 400, error };
+  const validation = validateTodoInput({
+    content,
+    description,
+    type_id,
+    isDone,
+  });
+  if (!validation.valid) return { status: 400, message: validation.message };
 
   if (type_id) {
-    const typeExists = await todoRepo.checkTypeIdExists(type_id);
+    const typeExists = await todoRepo.findTypeById(type_id);
     if (!typeExists) return { status: 400, message: "Type_id does not exist" };
   }
   try {
     await todoRepo.createTodo(content, description, type_id, isDone ?? 0);
     return { status: 201, message: "Todo created" };
   } catch (err) {
-    return { status: 500, error: "Database error" };
+    return { status: 500, message: "Database error" };
   }
 };
 
 exports.updateTodo = async (id, content, description, type_id, isDone) => {
-  const error = validateTodoInput({ content, description, type_id, isDone });
-  if (error) return { status: 400, error };
+  const validation = validateTodoInput({
+    content,
+    description,
+    type_id,
+    isDone,
+  });
+  if (!validation.valid) return { status: 400, message: validation.message };
 
   const todo = await todoRepo.getTodoById(id);
   if (!todo) {
     return { status: 404, message: "Todo not found" };
   }
   if (type_id) {
-    const typeExists = await todoRepo.checkTypeIdExists(type_id);
+    const typeExists = await todoRepo.findTypeById(type_id);
     if (!typeExists) {
       return { status: 400, message: "Type_id does not exist" };
     }
@@ -80,7 +63,7 @@ exports.updateTodo = async (id, content, description, type_id, isDone) => {
     await todoRepo.updateTodo(id, content, description, type_id, isDone);
     return { status: 200, message: "Todo updated" };
   } catch (err) {
-    return { status: 500, error: "Database error" };
+    return { status: 500, message: "Database error" };
   }
 };
 
@@ -93,6 +76,6 @@ exports.deleteTodo = async (id) => {
     await todoRepo.deleteTodo(id);
     return { status: 200, message: "Todo deleted" };
   } catch (err) {
-    return { status: 500, error: "Database error" };
+    return { status: 500, message: "Database error" };
   }
 };
